requirements.txt
asgiref==3.9.0
certifi==2025.6.15
charset-normalizer==3.4.2
Django==5.2.4
flake8==7.3.0
idna==3.10
jalali_core==1.0.0
jdatetime==5.0.0
mccabe==0.7.0
pillow==11.3.0
pycodestyle==2.14.0
pyflakes==3.4.0
python-dotenv==1.1.1
pytz==2025.2
requests==2.32.4
sqlparse==0.5.3
urllib3==2.5.0
